backend/app/requirements.txt
FastAPI
uvicorn
pydantic
sqlalchemy
redis
python-dotenv
whisper
deepgram-sdk
pytest
httpx
psycopg2-binary
kafka-python


